updated codes backend and frontend

+++ Backend_SpringBoot/CampusBackend/src/main/java/com/campusops/controller/PaymentController.java
@@ -4,6 +4,9 @@ import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import com.campusops.payment.PaymentService;
+import com.campusops.payment.PaymentVerifyRequest;
+
 @RestController
 @RequestMapping("/api/payment")
 @CrossOrigin(origins = "*")
@@ -13,7 +16,7 @@ public class PaymentController {
     private PaymentService paymentService;
 
     @PostMapping("/create-order/{registrationId}")
-    public String createOrder(@PathVariable Long registrationId) {
+    public String createOrder(@PathVariable int registrationId) {
 
         JSONObject order = paymentService.createOrder(registrationId);
         return order.toString();
